+++ frontend/libs/app-providers/src/observability/sentry/SentryErrorBoundary/index.tsx
