+++ frontend/libs/app-providers/src/observability/sentry/SentryErrorBoundary/style.ts
